코멘트 반영해 수정

+++ src/main/kotlin/com/blaybus/backend/entity/Task.kt
@@ -51,7 +51,7 @@ class Task(
     var writer: User,
 
     @Column(nullable = true)
-    var studyDurationInMinutes: Long? = null,
+    var studyDurationInMinutes: Int? = null,
 
     @Column(nullable = false)
     var isCompleted: Boolean = false,
